+++ src/manifolds/Grassmann.jl
@@ -90,7 +90,7 @@ end
     change_representer(M::Grassmann, ::EuclideanMetric, p, X)
 
 Change `X` to the corresponding representer of a cotangent vector at `p`.
-Since the [`Grassmann`](@ref) manifold `M`, is isometrically embedded, this is the identity
+Since the [`Grassmann`](@ref) manifold `M` is isometrically embedded, this is the identity.
 """
 change_representer(::Grassmann, ::EuclideanMetric, ::Any, ::Any)